Move service_metadata types to restate_common
