+++ src/common/src/lib.rs
@@ -2,5 +2,6 @@ pub mod errors;
 pub mod journal;
 pub mod partitioner;
 pub mod retry_policy;
+pub mod service_metadata;
 pub mod traits;
 pub mod types;
