+++ src/invoker/src/service/invocation_task.rs
@@ -17,12 +17,12 @@ use opentelemetry_http::HeaderInjector;
 use restate_common::errors::{InvocationError, InvocationErrorCode, UserErrorCode};
 use restate_common::journal::raw::{Header, PlainRawEntry, RawEntryHeader};
 use restate_common::journal::{Completion, EntryType};
+use restate_common::service_metadata::{EndpointMetadata, ProtocolType};
 use restate_common::types::{
     EnrichedRawEntry, EntryIndex, JournalMetadata, PartitionLeaderEpoch, ServiceInvocationId,
     ServiceInvocationSpanContext,
 };
 use restate_errors::warn_it;
-use restate_service_metadata::{EndpointMetadata, ProtocolType};
 use restate_service_protocol::message::{
     Decoder, Encoder, EncodingError, MessageHeader, MessageType, ProtocolMessage,
 };
