+++ src/invoker/src/service/state_machine_coordinator.rs
@@ -4,9 +4,10 @@ use crate::EntryEnricher;
 use codederror::CodedError;
 use restate_common::errors::UserErrorCode::Internal;
 use restate_common::journal::raw::Header;
+use restate_common::service_metadata::ProtocolType;
 use restate_common::types::EnrichedRawEntry;
 use restate_errors::warn_it;
-use restate_service_metadata::{ProtocolType, ServiceEndpointRegistry};
+use restate_service_metadata::ServiceEndpointRegistry;
 use std::collections::HashSet;
 use std::time::SystemTime;
 use tracing::{instrument, warn};
