+++ src/meta/src/rest_api/services.rs
@@ -6,7 +6,8 @@ use axum::Json;
 use hyper::http::{HeaderName, HeaderValue};
 use okapi_operation::*;
 use restate_common::retry_policy::RetryPolicy;
-use restate_service_metadata::{EndpointMetadata, ServiceEndpointRegistry};
+use restate_common::service_metadata::EndpointMetadata;
+use restate_service_metadata::ServiceEndpointRegistry;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use serde_with::serde_as;
