+++ src/meta/src/storage.rs
@@ -5,8 +5,8 @@ use futures::future::BoxFuture;
 use futures::stream::BoxStream;
 use futures::{stream, FutureExt, StreamExt};
 use prost_reflect::DescriptorPool;
+use restate_common::service_metadata::EndpointMetadata;
 use restate_service_key_extractor::ServiceInstanceType;
-use restate_service_metadata::EndpointMetadata;
 use serde::{Deserialize, Serialize};
 use tokio::fs::File;
 use tokio::io;
