+++ src/service_metadata/Cargo.toml
@@ -7,19 +7,7 @@ rust-version.workspace = true
 license.workspace = true
 publish = false
 
-[features]
-default = []
-serde = ["dep:serde", "dep:serde_with", "restate_common/serde", "restate_service_key_extractor/serde"]
-serde_schema = ["serde", "dep:schemars", "restate_common/serde_schema"]
-
 [dependencies]
-anyhow = { workspace = true }
 arc-swap = { workspace = true }
-base64 = { workspace = true }
 restate_common = { workspace = true }
-schemars = { workspace = true, optional = true }
-serde = { workspace = true, optional = true }
-serde_with = { workspace = true, optional = true }
-restate_service_key_extractor = { workspace = true }
-http = "0.2.9"
 prost-reflect = { workspace = true }
