+++ src/service_protocol/Cargo.toml
@@ -12,7 +12,7 @@ default = []
 protocol = []
 message = ["protocol", "dep:restate_common", "dep:bytes-utils", "dep:codederror", "dep:restate_errors", "dep:size", "dep:tracing"]
 codec = ["protocol", "dep:restate_common", "dep:thiserror", "dep:paste"]
-discovery = ["dep:thiserror", "dep:codederror", "dep:restate_errors", "dep:restate_service_key_extractor", "dep:hyper", "dep:hyper-rustls", "dep:prost-reflect", "dep:restate_common", "dep:restate_hyper_util", "dep:restate_service_metadata"]
+discovery = ["dep:thiserror", "dep:codederror", "dep:restate_errors", "dep:restate_service_key_extractor", "dep:hyper", "dep:hyper-rustls", "dep:prost-reflect", "dep:restate_common", "dep:restate_hyper_util"]
 
 [dependencies]
 bytes = { workspace = true }
@@ -30,7 +30,6 @@ hyper = { workspace = true, features = ["http1", "http2", "client", "tcp", "runt
 hyper-rustls = { workspace = true, optional = true }
 restate_common = { workspace = true, optional = true }
 prost-reflect = { workspace = true, optional = true }
-restate_service_metadata = { workspace = true, optional = true }
 codederror = { workspace = true, optional = true }
 restate_errors = { workspace = true, optional = true }
 restate_hyper_util = { workspace = true, optional = true }
