+++ src/service_protocol/src/discovery.rs
@@ -12,10 +12,10 @@ use prost_reflect::{
     ServiceDescriptor,
 };
 use restate_common::retry_policy::RetryPolicy;
+use restate_common::service_metadata::ProtocolType;
 use restate_errors::{META0001, META0002, META0003};
 use restate_hyper_util::proxy_connector::{Proxy, ProxyConnector};
 use restate_service_key_extractor::{KeyStructure, ServiceInstanceType};
-use restate_service_metadata::ProtocolType;
 
 // Clippy false positive, might be caused by Bytes contained within HeaderValue.
 // https://github.com/rust-lang/rust/issues/40543#issuecomment-1212981256
